pyproject.toml
[project]
name = "helloagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"

dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asgiref==3.8.1",
    "asyncstdlib==3.13",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "betterproto==2.0.0b7",
    "black==25.1.0",
    "blinker==1.9.0",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "chromadb>=1.3.5",
    "click==8.1.8",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "dataclasses-json==0.6.7",
    "datasets==3.6.0",
    "Deprecated==1.2.18",
    "dill==0.3.8",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "durationpy==0.9",
    "email_validator==2.2.0",
    "et_xmlfile==2.0.0",
    "fastapi==0.115.9",
    "fastapi-cli==0.0.7",
    "filelock==3.18.0",
    "fireworks-ai==0.16.3",
    "Flask==3.1.1",
    "flatbuffers==25.2.10",
    "frozenlist==1.6.0",
    "fsspec==2025.3.0",
    "google-auth==2.40.1",
    "google-search-results==2.4.2",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.3",
    "grpcio==1.71.0",
    "grpclib==0.4.8",
    "h11==0.16.0",
    "h2==4.2.0",
    "hf-xet==1.1.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "httpx-ws==0.7.2",
    "huggingface-hub==0.31.1",
    "humanfriendly==10.0",
    "hyperframe==6.1.0",
    "idna==3.10",
    "importlib_metadata==8.0.0",
    "importlib_resources==6.5.2",
    "isort==5.13.2",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "jiter>=0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "kubernetes==32.0.1",
    "langchain>=1.2.13",
    "langchain-tavily>=0.2.17",
    "langchain-community==0.4.1",
    "langchain-openai==1.1.11",
    "langchain-text-splitters==1.1.1",
    "langchain-chroma==1.1.0",
    "langchain-classic==1.0.3",
    "langchain-core==1.2.19",
    "langchainhub==0.1.21",
    "langchain-experimental==0.4.1",
    "langchain-fireworks==1.1.0",
    "langgraph==1.1.1",
    "langgraph-checkpoint>=2.1.0,<5.0.0",
    "langgraph-prebuilt>=1.0.8,<1.1.0",
    "langgraph-sdk>=0.3.0,<0.4.0",
    "langsmith>=0.3.45,<1.0.0",
    "llguidance==0.7.19",
    "markdown-it-py==3.0.0",
    "MarkupSafe==3.0.2",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "mmh3==5.1.0",
    "mpmath==1.3.0",
    "multidict==6.4.3",
    "multiprocess==0.70.16",
    "mypy_extensions==1.1.0",
    "numexpr==2.10.2",
    "numpy==2.2.5",
    "oauthlib==3.2.2",
    "onnxruntime==1.22.0",
    "openai>=2.0.0,<3.0.0",
    "openpyxl==3.1.5",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-exporter-otlp-proto-common>=1.40.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.40.0",
    "opentelemetry-instrumentation===0.61b0",
    "opentelemetry-instrumentation-asgi===0.61b0",
    "opentelemetry-instrumentation-fastapi===0.61b0",
    "opentelemetry-proto>=1.40.0",
    "opentelemetry-sdk==1.40.0",
    "opentelemetry-semantic-conventions===0.61b0",
    "opentelemetry-util-http===0.61b0",
    "orjson==3.10.18",
    "ormsgpack==1.10.0",
    "overrides==7.7.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pathspec==0.12.1",
    "pillow==11.2.1",
    "platformdirs==4.3.8",
    "posthog==4.0.1",
    "propcache==0.3.1",
    "protobuf>=5.29.3",
    "pyarrow==20.0.0",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.2",
    "pydantic==2.11.4",
    "pydantic-settings==2.10.1",
    "pydantic_core==2.33.2",
    "Pygments==2.19.1",
    "PyMySQL==1.1.1",
    "PyPika==0.48.9",
    "pyproject_hooks==1.2.0",
    "pyreadline3==3.5.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "PyYAML==6.0.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.0.0",
    "rich-toolkit==0.14.5",
    "rpds-py==0.24.0",
    "rsa==4.9.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.40",
    "starlette==0.45.3",
    "sympy==1.13.1",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "tqdm==4.67.1",
    "typer==0.15.3",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.0",
    "typing_extensions==4.13.2",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "Werkzeug==3.1.3",
    "wrapt==1.17.2",
    "wsproto==1.2.0",
    "xxhash==3.5.0",
    "yarl==1.20.0",
    "zipp==3.21.0",
    "zstandard==0.23.0",
    # multi_agent
    "aiofiles",
    "asyncio-atexit==1.0.1",
    "autogen-agentchat==0.7.5",
    "autogen-core==0.7.5",
    "autogen-ext==0.7.5",
    "docker==7.1.0",
    "jsonref==1.1.0",
]

